# api-schema/openapi.json
{
  "components": {
    "schemas": {
      "Agent": {
        "properties": {
          "alignment": {
            "title": "Alignment",
            "type": "string"
          },
          "alive": {
            "title": "Alive",
            "type": "boolean"
          },
          "attributes": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "string"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attributes"
          },
          "beliefs": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beliefs"
          },
          "emotions": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emotions"
          },
          "goal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "isHf": {
            "title": "Ishf",
            "type": "boolean"
          },
          "motivations": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motivations"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "needs": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Needs"
          },
          "pack": {
            "title": "Pack",
            "type": "string"
          },
          "persona": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona"
          },
          "recentEvents": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/WorldEvent"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recentevents"
          },
          "relationships": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Relationship"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationships"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "tile": {
            "title": "Tile",
            "type": "string"
          },
          "weeklyPlan": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weeklyplan"
          },
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "pack",
          "alignment",
          "isHf",
          "alive",
          "tile",
          "x",
          "y",
          "tags"
        ],
        "title": "Agent",
        "type": "object"
      },
      "BootstrapBody": {
        "properties": {
          "packs": {
            "items": {
              "type": "string"
            },
            "title": "Packs",
            "type": "array"
          },
          "seed": {
            "default": 42,
            "title": "Seed",
            "type": "integer"
          }
        },
        "required": [
          "packs"
        ],
        "title": "BootstrapBody",
        "type": "object"
      },
...
      "DiversitySummary": {
        "properties": {
          "top_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Kind"
          },
          "top_pct": {
            "title": "Top Pct",
            "type": "number"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "unique": {
            "title": "Unique",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "unique",
          "top_pct"
        ],
        "title": "DiversitySummary",
        "type": "object"
      },
      "EventKindCount": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "count"
        ],
        "title": "EventKindCount",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Relationship": {
        "properties": {
          "affinity": {
            "title": "Affinity",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "target",
          "kind",
          "affinity"
        ],
        "title": "Relationship",
        "type": "object"
      },
...
      "TickBody": {
        "properties": {
          "n": {
            "default": 1,
            "title": "N",
            "type": "integer"
          }
        },
        "title": "TickBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorldEvent": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "importance": {
            "title": "Importance",
            "type": "number"
          },
          "isEmergent": {
            "title": "Isemergent",
            "type": "boolean"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "narrative": {
            "title": "Narrative",
            "type": "string"
          },
          "outcome": {
            "enum": [
              "success",
              "failure",
              "neutral"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "pack": {
            "title": "Pack",
            "type": "string"
          },
          "participants": {
            "items": {
              "type": "string"
            },
            "title": "Participants",
            "type": "array"
          },
          "tick": {
            "title": "Tick",
            "type": "integer"
          },
          "tier": {
            "title": "Tier",
            "type": "integer"
          },
          "tile": {
            "title": "Tile",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tick",
          "pack",
          "tile",
          "kind",
          "outcome",
          "importance",
          "tier",
          "isEmergent",
          "participants",
          "narrative"
        ],
        "title": "WorldEvent",
        "type": "object"
      },
      "WorldSnapshot": {
        "properties": {
          "agentsAlive": {
            "title": "Agentsalive",
            "type": "integer"
          },
          "agentsTotal": {
            "title": "Agentstotal",
            "type": "integer"
          },
          "chapters": {
            "title": "Chapters",
            "type": "integer"
          },
          "deaths": {
            "title": "Deaths",
            "type": "integer"
          },
          "diversity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiversitySummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "eventsTotal": {
            "title": "Eventstotal",
            "type": "integer"
          },
          "loaded": {
            "title": "Loaded",
            "type": "boolean"
          },
          "modelTier2": {
            "title": "Modeltier2",
            "type": "string"
          },
          "modelTier3": {
            "title": "Modeltier3",
            "type": "string"
          },
          "packs": {
            "items": {
              "type": "string"
            },
            "title": "Packs",
            "type": "array"
          },
          "tick": {
            "title": "Tick",
            "type": "integer"
          },
          "tiles": {
            "title": "Tiles",
            "type": "integer"
          }
        },
        "required": [
          "loaded",
          "tick",
          "packs",
          "agentsAlive",
          "agentsTotal",
          "eventsTotal",
          "deaths",
          "chapters",
          "tiles",
          "modelTier2",
          "modelTier3"
        ],
        "title": "WorldSnapshot",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Living World API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agent/{agent_id}": {
      "get": {
        "operationId": "agent_api_agent__agent_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Agent"
      }
    },
...
    "/api/bootstrap": {
      "post": {
        "operationId": "bootstrap_api_bootstrap_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BootstrapBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bootstrap"
      }
    },
...
    "/api/event_kinds": {
      "get": {
        "operationId": "event_kinds_api_event_kinds_get",
        "parameters": [
          {
            "in": "query",
            "name": "top_k",
            "required": false,
            "schema": {
              "default": 10,
              "title": "Top K",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventKindCount"
                  },
                  "title": "Response Event Kinds Api Event Kinds Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event Kinds"
      }
    },
    "/api/events": {
      "get": {
        "operationId": "events_api_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Since",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 80,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorldEvent"
                  },
                  "title": "Response Events Api Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events"
      }
    },
...
    "/api/settings": {
      "get": {
        "description": "Settings shape is deep + dynamic — exposed as open dict for now.\nTyping it is a follow-up (one Pydantic model per section).",
        "operationId": "settings_get_api_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Settings Get Api Settings Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Settings Get"
      },
      "post": {
        "description": "Patch the live settings + persist to settings.yaml.\n\nAccepts a partial dict — only top-level sections present in the\npayload are merged. Returns the merged settings dict.\n\nNOTE: Most settings only take effect on next bootstrap (engine\nis built from settings at bootstrap time). Caller should\nprompt the user to Reset → Simulate again after important\nchanges (model swap, packs, feature flags).",
        "operationId": "settings_set_api_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Settings Set Api Settings Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Settings Set"
      }
    },
...
    "/api/tick": {
      "post": {
        "operationId": "tick_api_tick_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TickBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorldSnapshot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tick"
      }
    },
...
  }
}